backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.1
decorator==5.1.1
ecdsa==0.18.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
mysql-connector-python==8.3.0
passlib==1.7.4
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
pydantic-sqlalchemy==0.0.9
pydantic_core==2.16.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
retry==0.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
